pyproject.toml
[project]
name = "victury-group-test-task"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=1.4",
    "asyncpg",
    "aiogram",
    "alembic",
    "apscheduler",
    "redis",
    "python-decouple",
    "python-logging-loki",
    "pytz",
]

[dependency-groups]
dev = [
    "ruff",
]

[tool.ruff]
line-length = 88

target-version = "py312"

exclude = [
    ".git",
    "__pycache__",
    "alembic/versions/*.py",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C90",
]


ignore = [
    "E501",
    "B008",
    "C901",
]

[tool.ruff.format]

quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
